+++ warp/controller/utils_test.go
@@ -1,0 +1,14 @@
+package controller
+
+type WarpMenuCategory struct {
+	Title   string
+	Order   int
+	Entries []WarpMenuEntry
+}
+
+type WarpMenuEntry struct {
+	Title       string
+	DisplayName string
+	Href        string
+	Target      string
+}